+++ javascript/core/sparkplug-payload/lib/sparkplugbpayload.ts
@@ -141,10 +141,10 @@ function setValue (type: number, value: UserValue, object: IMetric | IPropertyVa
             break;
         case 4: // Int64
             if (value as number >= 0) {
-                object.intValue = value as number;
+                object.longValue = value as number;
                 break;
             }
-            object.intValue = value as number + 2 ** 64;
+            object.longValue = value as number + 2 ** 64;
             break;
         case 7: // UInt32
         case 8: // UInt64
